Rename spark30 to spark300
