+++ shims/pom.xml
@@ -26,16 +26,16 @@
         <relativePath>../pom.xml</relativePath>
     </parent>
     <groupId>com.nvidia</groupId>
-    <artifactId>rapids-4-spark-shims-aggregator_2.12</artifactId>
+    <artifactId>rapids-4-spark-shims_2.12</artifactId>
     <packaging>pom</packaging>
     <name>RAPIDS Accelerator for Apache Spark SQL Plugin Shims</name>
     <description>The RAPIDS SQL plugin for Apache Spark Shims</description>
     <version>0.2.0-SNAPSHOT</version>
 
     <modules>
-        <module>spark30</module>
+        <module>spark300</module>
         <module>spark301</module>
-        <module>spark31</module>
+        <module>spark310</module>
         <module>aggregator</module>
     </modules>
     <dependencies>
@@ -43,11 +43,13 @@
             <groupId>com.nvidia</groupId>
             <artifactId>rapids-4-spark-sql_${scala.binary.version}</artifactId>
             <version>${project.version}</version>
+            <scope>provided</scope>
         </dependency>
         <dependency>
             <groupId>ai.rapids</groupId>
             <artifactId>cudf</artifactId>
             <classifier>${cuda.version}</classifier>
+            <scope>provided</scope>
         </dependency>
     </dependencies>
     <build>
